LRQA-41876: Update portlet hub tests
