+++ modules/apps/frontend-js/frontend-js-web/src/main/resources/META-INF/resources/liferay/portlet/PortletInit.es.js
@@ -194,7 +194,7 @@ class PortletInit {
 									resolve(updatedIds);
 								}
 							)
-							.catch (
+							.catch(
 								err => {
 									reject(err);
 								}
@@ -578,7 +578,7 @@ class PortletInit {
 					this._updatePortletStates(updatedIds);
 				}
 			)
-			.catch (
+			.catch(
 				err => {
 					busy = false;
 					this._reportError(this._portletId, err);
@@ -695,7 +695,7 @@ class PortletInit {
 						resolve(val);
 					}
 				)
-				.catch (
+				.catch(
 					err => {
 						reject(err);
 					}
